OS Err 5
Getting OS error code 5: Access Denied when I try to run this code. I believe it is happening because I cannot kill the thread correctly? Perhaps? I am not sure, but I already tried changing my windows defender exceptions to include my Capstone code folder. I will attempt a fix tomorrow
