+++ relay/src/state.rs
@@ -6,7 +6,7 @@ use iced::time::Duration;
 #[allow(unused)]
 pub(crate) struct State {
     pub elapsed_time: Duration,
-    pub thread_handle: Option<JoinHandle<()>>,
+    pub ipc_conn_thread_handle: Option<JoinHandle<()>>,
     pub tx_kill: Option<std::sync::mpsc::Sender<()>>,
     pub rx_baton: Option<std::sync::mpsc::Receiver<f32>>,
     pub latest_baton_send: Option<f32>,
