+++ relay/src/update.rs
@@ -1,6 +1,6 @@
 use iced::{time::Duration, Task};
 
-use crate::{Message, State};
+use crate::{Message, State, ipc::ipc_connection_loop};
 
 pub(crate) fn update(state: &mut State, message: Message) -> Task<Message> {
     use Message as M;
@@ -37,6 +37,27 @@ pub(crate) fn update(state: &mut State, message: Message) -> Task<Message> {
             if let Some(num) = state.rx_baton.as_ref().and_then(|rx| rx.try_recv().ok()) {
                 state.latest_baton_send = Some(num);
             }
+            Task::none()
+        }
+        M::BatonReconnectMessage => {
+            println!("Baton Reconnect Message");
+            // Kill existing IPC thread
+            if let Some(tx_kill) = &state.tx_kill {
+                let _ = tx_kill.send(()); 
+            }
+
+            // Create new IPC thread
+            let (tx_kill, rx_kill) = std::sync::mpsc::channel();
+            let (txx, rxx) = std::sync::mpsc::channel();
+            let handle = std::thread::spawn(move || {
+                ipc_connection_loop(rx_kill, txx);
+            });
+
+            // Update state with new IPC thread
+            state.ipc_conn_thread_handle = Some(handle);
+            state.tx_kill = Some(tx_kill);
+            state.rx_baton = Some(rxx);
+
             Task::none()
         }
         _ => Task::none(),
